Add files via upload
Part 1: PCA:

Problem Statement: 
The ‘Hair Salon.csv’  dataset contains various variables used for the context of Market Segmentation. This particular case study is based on various parameters of a salon chain of hair products. You are expected to do a Principal Component Analysis for this case study according to the instructions given in the rubric. Kindly refer to the PCA_Data_Dictionary.jpg  file for the Data Dictionary of the Dataset.  
Note: This particular dataset contains the target variable satisfaction as well. Please do drop this variable before doing Principal Component Analysis.

Part 2: Clustering:

The  State_wise_Health_income.csv  dataset given is about the Health and economic conditions in different States of a country. The Group States based on how similar their situation is, so as to provide these groups to the government so that appropriate measures can be taken to escalate their Health and Economic conditions.

- Read the data and do exploratory data analysis. Describe the data briefly. (Check the null values, Data types, shape, EDA, etc, etc)
- Do you think scaling is necessary for clustering in this case? Justify
- Apply hierarchical clustering to scaled data. Identify the number of optimum clusters using Dendrogram and briefly describe them.
- Apply K-Means clustering on scaled data and determine optimum clusters. Apply elbow curve and find the silhouette score.
-  Describe cluster profiles for the clusters defined. Recommend different priority based actions that need to be taken for different clusters on the bases of their vulnerability situations according to their Economic and Health Conditions.

Data Dictionary for State_wise_Health_income Dataset:

- States: names of States
- Health_indeces1: A composite index rolls several related measures (indicators) into a single score that provides a summary of how the health system is performing in the State.
- Health_indeces2: A composite index rolls several related measures (indicators) into a single score that provides a summary of how the health system is performing in certain areas of the State.
- Per_capita_income-Per capita income (PCI) measures the average income earned per person in a given area (city, region, country, etc.) in a specified year. It is calculated by dividing the area's total income by its total population.
- GDP: GDP provides an economic snapshot of a country/state, used to estimate the size of an economy and its growth rate.